creation anonyme user all user can delete anonyme task

+++ src/Controller/TaskController.php
@@ -86,10 +86,12 @@ class TaskController extends AbstractController
      */
     public function delete(Request $request, Task $task): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$task->getId(), $request->request->get('_token'))) {
+        if ($this->getUser() === $task->getUser() && $this->isCsrfTokenValid('delete'.$task->getId(), $request->request->get('_token'))) {
+
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($task);
             $entityManager->flush();
+
         }
 
         return $this->redirectToRoute('task_list', [], Response::HTTP_SEE_OTHER);
